requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
Django==3.2.6
django-appconf==1.0.5
django-environ==0.9.0
django-filter==22.1
django-imagekit==4.1.0
djangorestframework==3.13.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==4.12.0
Markdown==3.4.1
pilkit==2.0
Pillow==9.2.0
promise==2.3
psycopg2-binary==2.9.3
pytz==2022.2.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
zipp==3.8.1
